+++ lib/src/cli/commands/extract_command.dart
@@ -9,10 +9,10 @@ import 'command_mixin.dart';
 String _optionNameInput = 'input';
 String _optionNameIncludePathDependencies = 'include-path-dependencies';
 String _optionNameOutput = 'output';
-String _optionNameNoMergeBaseClasses = 'no-merge-base-classes';
 String _optionNameNoAnalyzePlatformConstraints =
     'no-analyze-platform-constraints';
 String _optionNameRemoveExample = 'remove-example';
+String _optionNameSetExitCodeOnMissingExport = 'set-exit-on-missing-export';
 
 /// command to extract the public API of a package.
 /// This is used when, for example, the public API needs to be stored on disk
@@ -42,12 +42,6 @@ If not specified the extracted API will be printed to the console.
 ''',
     );
     argParser.addFlag(
-      _optionNameNoMergeBaseClasses,
-      help: 'Disables base class merging.',
-      defaultsTo: false,
-      negatable: false,
-    );
-    argParser.addFlag(
       _optionNameNoAnalyzePlatformConstraints,
       help: 'Disables analysis of platform constraints.',
       defaultsTo: false,
@@ -59,6 +53,13 @@ If not specified the extracted API will be printed to the console.
       defaultsTo: true,
       negatable: true,
     );
+    argParser.addFlag(
+      _optionNameSetExitCodeOnMissingExport,
+      help:
+          'Sets exit code to != 0 if missing exports are detected in the API.',
+      defaultsTo: false,
+      negatable: true,
+    );
   }
 
   @override
@@ -66,11 +67,11 @@ If not specified the extracted API will be printed to the console.
     final packageRef = PackageRef(argResults![_optionNameInput]);
     final shouldCheckPathDependencies =
         argResults![_optionNameIncludePathDependencies] as bool;
-    final noMergeBaseClasses =
-        argResults![_optionNameNoMergeBaseClasses] as bool;
     final noAnalyzePlatformConstraints =
         argResults![_optionNameNoAnalyzePlatformConstraints] as bool;
     final doRemoveExample = argResults![_optionNameRemoveExample] as bool;
+    final doSetExitCodeOnMissingExport =
+        argResults![_optionNameSetExitCodeOnMissingExport] as bool;
 
     final preparedPackageRef = await prepare(
       packageRef,
@@ -78,7 +79,6 @@ If not specified the extracted API will be printed to the console.
     );
     final packageApi = await analyze(
       preparedPackageRef,
-      doMergeBaseClasses: !noMergeBaseClasses,
       doAnalyzePlatformConstraints: !noAnalyzePlatformConstraints,
       doRemoveExample: doRemoveExample,
     );
@@ -95,6 +95,19 @@ If not specified the extracted API will be printed to the console.
       stdout.writeln('Public API of "$packageRef" written to $outFilePath');
     } else {
       stdout.writeln(jsonString);
+    }
+
+    final declarationsWithoutEntryPoints =
+        packageApi.getRootDeclarationsWithoutEntryPoints();
+    if (declarationsWithoutEntryPoints.isNotEmpty) {
+      stdout.writeln(
+          'The following declarations do not have an entry point (did you miss to export them?):');
+      for (final declaration in declarationsWithoutEntryPoints) {
+        stdout.writeln('  ${declaration.name}');
+      }
+      if (doSetExitCodeOnMissingExport) {
+        return -1;
+      }
     }
     return 0;
   }